Make RowRenderer a separate file
